refactor: :truck: Organizei melhor telas de crud de conta bancária.
