+++ frontend/src/features/conta-bancaria/ui/steps/step-detalhe.tsx
